getting rid of loggerutil package, moving this stuff to logger
